Move SafeMaskMaker to gammapy/makers
