+++ gammapy/analysis/core.py
@@ -5,7 +5,8 @@ from astropy.coordinates import SkyCoord
 from astropy.table import Table
 from regions import CircleSkyRegion
 from gammapy.analysis.config import AnalysisConfig
-from gammapy.cube import FoVBackgroundMaker, MapDatasetMaker, SafeMaskMaker
+from gammapy.cube import FoVBackgroundMaker, MapDatasetMaker
+from gammapy.makers import SafeMaskMaker
 from gammapy.datasets import MapDataset, SpectrumDataset, FluxPointsDataset
 from gammapy.data import DataStore
 from gammapy.maps import Map, MapAxis, WcsGeom
