+++ gammapy/makers/__init__.py
@@ -1,1 +1,2 @@
 from .ring import *
+from .safe import *
